requirements.txt
absl-py==0.11.0
astunparse==1.6.3
bert-extractive-summarizer==0.6.1
blis==0.4.1
boto3==1.16.63
botocore==1.19.63
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.2.0/en_core_web_lg-2.2.0.tar.gz
filelock==3.0.12
flatbuffers==1.12
fpdf==1.7.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
murmurhash==1.0.5
neuralcoref==4.0
nltk==3.5
numpy==1.20.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pathy==0.3.4
pdfminer.six==20200517
pdfplumber==0.5.25
Pillow==8.1.0
plac==0.9.6
preshed==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.9
pydantic==1.7.3
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==2.2.0
srsly==1.0.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==7.1.1
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
tqdm==4.56.0
transformers==4.2.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Wand==0.6.5
wasabi==0.8.2
Werkzeug==1.0.1
wrapt==1.12.1
